Create separate controller for user auth purposes. Create basic login/signup pages.
