+++ src/main/resources/templates/index.html
@@ -20,7 +20,7 @@
     </script>
 </head>
 <body>
-<nav x-data="{ open: false, dropdownOpen: false }" class="bg-white shadow">
+<nav x-data="{ open: false, dropdownOpen: false }" class="bg-white">
     <div class="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
         <div class="flex h-16">
             <div class="flex">
@@ -50,8 +50,8 @@
                             Discover a fitness platform that allows you to craft a personalized workout regimen. Log daily exercises, view previous workouts, and add exercises.
                         </p>
                         <div class="mt-10 flex items-center gap-x-6">
-                            <a href="#" class="rounded-md bg-deep-red px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-lighter-red focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-deep-red">Sign Up</a>
-                            <a href="#" class="text-sm font-semibold leading-6 text-gray-900">Login <span aria-hidden="true">→</span></a>
+                            <a th:href="@{/signup}" class="rounded-md bg-deep-red px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-lighter-red focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-deep-red">Sign Up</a>
+                            <a th:href="@{/login}" class="text-sm font-semibold leading-6 text-gray-900">Log in &rarr;</a>
                         </div>
                     </div>
                 </div>
