+++ src/main/resources/templates/login.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang="en" class="h-full bg-white">
+<head>
+    <meta charset="UTF-8">
+    <title>Workout Builder — Log In</title>
+    <link rel="stylesheet" th:href="@{/css/application.css}">
+    <script src="//unpkg.com/alpinejs" defer></script>
+    <script src="https://cdn.tailwindcss.com"></script>
+    <script>
+        tailwind.config = {
+            theme: {
+                extend: {
+                    colors: {
+                        'deep-red': '#8B0000',
+                        'lighter-red': '#A10A0A',
+                    }
+                }
+            }
+        }
+    </script>
+</head>
+<body class="h-full">
+    <div class="flex min-h-full flex-col justify-center px-6 py-12 lg:px-8">
+        <div class="sm:mx-auto sm:w-full sm:max-w-sm">
+            <img class="mx-auto h-10 w-auto" th:src="@{/images/logo.svg}" alt="Workout Builder">
+            <h2 class="mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900">Login to Workout Builder</h2>
+        </div>
+        <div class="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
+            <form class="space-y-6" action="#" method="post">
+                <div>
+                    <label for="email" class="block text-sm font-medium leading-6 text-gray-900">Email address</label>
+                    <div class="mt-2">
+                        <input id="email" name="email" type="email" autocomplete="email" required class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-deep-red sm:text-sm sm:leading-6">
+                    </div>
+                </div>
+                <div>
+                    <div class="flex items-center justify-between">
+                        <label for="password" class="block text-sm font-medium leading-6 text-gray-900">Password</label>
+                    </div>
+                    <div class="mt-2">
+                        <input id="password" name="password" type="password" autocomplete="current-password" required class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-deep-red sm:text-sm sm:leading-6">
+                    </div>
+                </div>
+                <div>
+                    <button type="submit" class="flex w-full justify-center rounded-md bg-deep-red px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-lighter-red focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-deep-red">Log In</button>
+                </div>
+            </form>
+            <div class="mt-10 text-center text-sm text-gray-500 flex flex-col">
+                <span>Not a member? <a th:href="@{/signup}" class="font-semibold leading-6 text-deep-red hover:text-lighter-red">Sign Up &rarr;</a></span>
+                <a class="mt-8 text-deep-red hover:text-lighter-red" th:href="@{/}">Return to home</a>
+            </div>
+        </div>
+    </div>
+</body>
+</html>
